# codex_poting/golang/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 書籍管理システム
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  "/api/books/{id}/sales-unit-prices":
    post:
      tags:
      - Books
      summary: 本販売単価登録
      operationId: createSalesUnitPrice
      parameters:
      - name: id
        in: path
        description: 本ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BookSalesUnitPriceCreateRequest"
        required: true
      responses:
        '200':
          description: 成功
        '400':
          description: リクエストエラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                invalidRequestBody:
                  summary: リクエストボディのバリデーションエラー
                  description: invalidRequestBody
                  value:
                    errors:
                    - field: salesUnitPrice
                      message: 1 以上の値にしてください
                    - field: effectiveFrom
                      message: 未来の日付にしてください
                    instance: "/api/books/1/sales-unit-prices"
                    status: 400
                    title: リクエストバリデーションエラー
                salesUnitPriceEffectiveFromUniqueConstraintViolation:
                  summary: 同一書籍・同一有効開始日の一意制約違反
                  description: salesUnitPriceEffectiveFromUniqueConstraintViolation
                  value:
                    detail: '一意制約に違反しています: book_sales_unit_price_history(book_id,effective_from=1,2026-08-01)'
                    instance: "/api/books/1/sales-unit-prices"
                    status: 400
                    title: データバリデーション
        '401':
          description: 認証エラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                unauthorized:
                  summary: 認証トークンなし
                  description: unauthorized
                  value:
                    detail: Unauthorized
                    instance: "/api/books/1/sales-unit-prices"
                    status: 401
                    title: Unauthorized
        '404':
          description: データなし
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                bookNotFound:
                  summary: 本販売単価登録対象の本が存在しない
                  description: bookNotFound
                  value:
                    instance: "/api/books/999/sales-unit-prices"
                    status: 404
                    title: 該当データなし
        '409':
          description: 更新競合
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                updateConflict:
                  summary: 他ユーザーによる更新競合
                  description: updateConflict
                  value:
                    detail: 他ユーザーによって更新されています
                    instance: "/api/books/1/sales-unit-prices"
                    status: 409
                    title: 更新競合
      security:
      - bearerAuth: []
  "/api/books/update":
    post:
      tags:
      - Books
      summary: 本更新
      operationId: updateBook
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BookUpdateRequest"
        required: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookResponse"
        '400':
          description: リクエストエラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                invalidRequestBody:
                  summary: リクエストボディのバリデーションエラー
                  description: invalidRequestBody
                  value:
                    errors:
                    - field: title
                      message: 1 から 100 の間のサイズにしてください
                    - field: releaseDate
                      message: null は許可されていません
                    - field: isbn
                      message: 13桁の数字にしてください
                    instance: "/api/books/update"
                    status: 400
                    title: リクエストバリデーションエラー
                publisherReferenceNotFound:
                  summary: 出版社の参照先なし
                  description: publisherReferenceNotFound
                  value:
                    detail: '参照先データが存在しません: publisher(id=999)'
                    instance: "/api/books/update"
                    status: 400
                    title: データバリデーション
                isbnUniqueConstraintViolation:
                  summary: ISBNの一意制約違反
                  description: isbnUniqueConstraintViolation
                  value:
                    detail: '一意制約に違反しています: book(isbn=0000000000001)'
                    instance: "/api/books/update"
                    status: 400
                    title: データバリデーション
        '404':
          description: データなし
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                bookNotFound:
                  summary: 更新対象の本が存在しない
                  description: bookNotFound
                  value:
                    instance: "/api/books/update"
                    status: 404
                    title: 該当データなし
        '401':
          description: 認証エラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                unauthorized:
                  summary: 認証トークンなし
                  description: unauthorized
                  value:
                    detail: Unauthorized
                    instance: "/api/books/update"
                    status: 401
                    title: Unauthorized
        '409':
          description: 更新競合
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                updateConflict:
                  summary: 他ユーザーによる更新競合
                  description: updateConflict
                  value:
                    detail: 他ユーザーによって更新されています
                    instance: "/api/books/update"
                    status: 409
                    title: 更新競合
      security:
      - bearerAuth: []
  "/api/books/create":
    post:
      tags:
      - Books
      summary: 本登録
      operationId: createBook
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BookCreateRequest"
        required: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookResponse"
        '400':
          description: リクエストエラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                invalidRequestBody:
                  summary: リクエストボディのバリデーションエラー
                  description: invalidRequestBody
                  value:
                    errors:
                    - field: title
                      message: 1 から 100 の間のサイズにしてください
                    - field: releaseDate
                      message: null は許可されていません
                    - field: isbn
                      message: 13桁の数字にしてください
                    instance: "/api/books/create"
                    status: 400
                    title: リクエストバリデーションエラー
                publisherReferenceNotFound:
                  summary: 出版社の参照先なし
                  description: publisherReferenceNotFound
                  value:
                    detail: '参照先データが存在しません: publisher(id=999)'
                    instance: "/api/books/create"
                    status: 400
                    title: データバリデーション
                isbnUniqueConstraintViolation:
                  summary: ISBNの一意制約違反
                  description: isbnUniqueConstraintViolation
                  value:
                    detail: '一意制約に違反しています: book(isbn=0000000000001)'
                    instance: "/api/books/create"
                    status: 400
                    title: データバリデーション
        '401':
          description: 認証エラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                unauthorized:
                  summary: 認証トークンなし
                  description: unauthorized
                  value:
                    detail: Unauthorized
                    instance: "/api/books/create"
                    status: 401
                    title: Unauthorized
      security:
      - bearerAuth: []
...
  "/api/auth/login-rate-limit/reset":
    post:
      tags:
      - Auth
      summary: ログインリクエスト回数制限のリセット
      operationId: resetLoginRateLimit
      responses:
        '204':
          description: リセット成功
        '401':
          description: 認証エラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
  "/api/books/{id}":
    get:
      tags:
      - Books
      summary: 本取得
      operationId: getBook
      parameters:
      - name: id
        in: path
        description: 本ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookResponse"
        '400':
          description: リクエストエラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                invalidBookId:
                  summary: 本IDの形式不正
                  description: invalidBookId
                  value:
                    detail: 'Failed to convert ''id'' with value: ''abc'''
                    instance: "/api/books/abc"
                    status: 400
                    title: Bad Request
        '404':
          description: データなし
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                bookNotFound:
                  summary: 取得対象の本が存在しない
                  description: bookNotFound
                  value:
                    instance: "/api/books/999"
                    status: 404
                    title: 該当データなし
    delete:
      tags:
      - Books
      summary: 本削除
      operationId: deleteBook
      parameters:
      - name: id
        in: path
        description: 本ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: 成功
        '400':
          description: リクエストエラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                invalidBookId:
                  summary: 本IDの形式不正
                  description: invalidBookId
                  value:
                    detail: 'Failed to convert ''id'' with value: ''abc'''
                    instance: "/api/books/abc"
                    status: 400
                    title: Bad Request
        '404':
          description: データなし
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                bookNotFound:
                  summary: 削除対象の本が存在しない
                  description: bookNotFound
                  value:
                    instance: "/api/books/999"
                    status: 404
                    title: 該当データなし
        '401':
          description: 認証エラー
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                unauthorized:
                  summary: 認証トークンなし
                  description: unauthorized
                  value:
                    detail: Unauthorized
                    instance: "/api/books/1"
                    status: 401
                    title: Unauthorized
        '409':
          description: 更新競合
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/ProblemDetail"
              examples:
                updateConflict:
                  summary: 削除対象が他ユーザーにより更新中
                  description: updateConflict
                  value:
                    detail: 他ユーザーによって更新されています
                    instance: "/api/books/1"
                    status: 409
                    title: 更新競合
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
    BookSalesUnitPriceCreateRequest:
      type: object
      properties:
        salesUnitPrice:
          type: string
          description: 販売単価
          example: 1500
          maximum: 10000
          minimum: 1
        effectiveFrom:
          type: string
          format: date
          description: 有効開始日
      required:
      - effectiveFrom
      - salesUnitPrice
    BookUpdateRequest:
      type: object
      properties:
        id:
          type: string
          format: int64
          description: 本ID
        title:
          type: string
          description: タイトル
          maxLength: 100
          minLength: 1
        author:
          type: string
          description: 著者
          maxLength: 200
          minLength: 0
        releaseDate:
          type: string
          format: date
          description: 発売日付
        publisherId:
          type: string
          format: int64
          description: 出版社ID
        genreId:
          type: string
          format: int64
          description: ジャンルID
        isbn:
          type: string
          description: ISBN
          example: '9784000000000'
        version:
          type: string
          format: int64
          description: バージョン
      required:
      - genreId
      - id
      - isbn
      - publisherId
      - releaseDate
      - title
      - version
    BookResponse:
      type: object
      properties:
        id:
          type: string
          format: int64
          description: 本ID
        title:
          type: string
          description: タイトル
        author:
          type: string
          description: 著者
        releaseDate:
          type: string
          format: date
          description: 発売日付
        publisherId:
          type: string
          format: int64
          description: 出版社ID
        publisherName:
          type: string
          description: 出版社名
        genreId:
          type: string
          format: int64
          description: ジャンルID
        genreName:
          type: string
          description: ジャンル名
        isbn:
          type: string
          description: ISBN
        salesUnitPrice:
          type: string
          description: 販売単価
        updateAt:
          type: string
          format: date-time
          description: 更新日時
        version:
          type: string
          format: int64
          description: バージョン
        bookStockList:
          type: array
          description: 本在庫リスト
          items:
            "$ref": "#/components/schemas/BookStockResponse"
      required:
      - bookStockList
      - genreId
      - genreName
      - id
      - isbn
      - publisherId
      - publisherName
      - releaseDate
      - salesUnitPrice
      - title
      - updateAt
      - version
    BookStockResponse:
      type: object
      properties:
        id:
          type: string
          format: int64
          description: 本在庫ID
        bookStockStoreId:
          type: string
          format: int64
          description: 本在庫店舗ID
        storeName:
          type: string
          description: 本在庫店舗名
        bookStockQuantity:
          type: string
          description: 本在庫数量
      required:
      - bookStockQuantity
      - bookStockStoreId
      - id
      - storeName
    BookCreateRequest:
      type: object
      properties:
        title:
          type: string
          description: タイトル
          maxLength: 100
          minLength: 1
        author:
          type: string
          description: 著者
          maxLength: 200
          minLength: 0
        releaseDate:
          type: string
          format: date
          description: 発売日付
        publisherId:
          type: string
          format: int64
          description: 出版社ID
        genreId:
          type: string
          format: int64
          description: ジャンルID
        isbn:
          type: string
          description: ISBN
          example: '9784000000000'
        salesUnitPrice:
          type: string
          description: 販売単価
          example: 1200
          maximum: 10000
          minimum: 1
      required:
      - genreId
      - isbn
      - publisherId
      - releaseDate
      - salesUnitPrice
      - title
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT